Move attester functions to the separate file
Ref: https://github.com/hl7au/au-ps-inferno/issues/43
